Update for greatexpectations integrations in DVC pipelinening and update in model class declaration
