+++ .dvc/config
@@ -1,5 +1,6 @@
 [core]
     remote = origin
+	autostage = True
 ['remote "origin"']
     url = https://dagshub.com/armand-07/TAED2-Falsettos.dvc
     read_timeout = 10000
