+++ data/raw_sample_example.dvc
@@ -1,5 +1,5 @@
 outs:
-- md5: d751713988987e9331980363e24189ce.dir
-  size: 0
-  nfiles: 0
+- md5: cee16202c3db652d6a6860d4e241fa7d.dir
+  size: 110020
+  nfiles: 1
   path: raw_sample_example
