+++ params.yaml
@@ -13,7 +13,7 @@ model:
   optimizer: 'adam' # optimization method: sgd | adam
   adapter_hidden_size: 128 # model hyperparameter
 
-  epochs: 100  # maximum number of epochs to train
+  epochs: 2  # maximum number of epochs to train
   patience: 5 # how many epochs of no loss improvement should we wait before stop training
   log_interval: 15 # how many batches to wait before logging training status
