+++ src/models/test_model.py
@@ -16,6 +16,7 @@ from torcheval.metrics.functional import multiclass_f1_score
 import torch.nn.functional as F
 
 from src import ROOT_DIR, MODELS_DIR, PROCESSED_DATA_DIR
+from src.models.Hubert_Classifier_model import HubertForAudioClassification
 
 from pathlib import Path
 
@@ -42,52 +43,6 @@ def data_loading(input_folder_path, batch_size):
     print('Test set has {} instances'.format(len(audio_dataset["test"])))
 
     return test_loader
-
-
-class HubertForAudioClassification(nn.Module):
-    def __init__(self, adapter_hidden_size = 64):
-        super().__init__()
-
-        self.hubert = HubertModel.from_pretrained("facebook/hubert-base-ls960")
-
-        hidden_size = self.hubert.config.hidden_size
-
-        self.adaptor = nn.Sequential(
-            nn.Linear(hidden_size, adapter_hidden_size),
-            nn.ReLU(True),
-            nn.Dropout(0.05),
-            nn.Linear(adapter_hidden_size, hidden_size),
-        )
-
-        self.classifier = nn.Sequential(
-            nn.Linear(hidden_size, adapter_hidden_size),
-            nn.ReLU(True),
-            nn.Dropout(0.05),
-            nn.Linear(adapter_hidden_size, 36), # 36 classes
-        )
-
-    def freeze_feature_encoder(self):
-        """
-        Calling this function will disable the gradient computation for the feature encoder so that its parameter will
-        not be updated during training.
-        """
-        self.hubert.feature_extractor._freeze_parameters()
-
-    def forward(self, x):
-        # x shape: (B,E)
-        x = self.hubert(x).last_hidden_state
-        x = F.layer_norm(x, x.shape[1:])
-        x = self.adaptor(x)
-        # pooling
-        x, _ = x.max(dim=1)
-
-        # Mutilayer perceptron with log softmax for classification
-        out = self.classifier(x)
-
-        # Remove last dimension
-        return out
-        # return shape: (B, total_labels)
-
 
 def test(loader, model, criterion, num_classes):
     model.eval()
